Add empty IComponentStore implementation
